Wrap expressions in `r.expr` except when we don't.
The logic here is sort of twisted; on the top level, JavaScript values
like strings and numbers are not ReQL datums and so do not accept any
of the usual arguments.  However arguments to ReQL functions are
automatically coerced to datums, and a nested datum like an array or
an object is converted wholesale.  Additionally for added irritation
the syntax for creating an object at the toplevel is different;
`r.object` instead of `r.expr({})`.