A sketch of what would be a Twisted driver I've didn't tested it yet, I don't have any RethinkDB config running as now (working on Windows at time of writing...)
Theorically, it should work, except how the timeout are handled. I should
dig a little more to see how connectTCP deal with timeout, and to see if I
can timeout a defer. In brief, there is some work to do to ensure that all
is okay.

(Fixed some strange things in the two others async drivers. And some
typo!)