Allow coercing arbitrary sequence types to objects
The code was limiting the source type in SEQUENCE -> OBJECT coercions to only be
an array, but this is misleading and unhelpful.

There's a possibility of using too much memory for this (because we have
an array size limit but no object size limit), but there are other solutions
being worked on for that problem more generally.

Fixes #2802