Fixes #6296 InvalidStateError after cancelling a RethinkDB operation.
If a Python coroutine is cancelled while it is suspended inside a RethinkDB query, the query
future is cancelled. When the connection is later closed, the `close()` method tries to set the
future result, which throws `InvalidStateError` because the future's was already cancelled. This
commit adds a guard condition to avoid this.